+++ .eslintrc.cjs
@@ -9,28 +9,31 @@ module.exports = {
     polyfills: [
     ]
   },
-  extends: ['ash-nazg/sauron-node-script-overrides'],
+  extends: ['ash-nazg/sauron-node-overrides'],
   parserOptions: {
-    ecmaVersion: 2018
+    ecmaVersion: 2022
   },
   overrides: [
     {
       files: '*.md/*.js',
-      extends: ['ash-nazg/sauron-node-script-overrides'],
+      extends: ['ash-nazg/sauron-node-overrides'],
       globals: {
         require: 'readonly',
         license: 'readonly',
         licenseBadger: 'readonly'
       },
       rules: {
-        'node/no-missing-require': ['error', {
+        'import/no-unresolved': ['error', {
+          ignore: ['license-badger/src/getLicenseType\\.js']
+        }],
+        'n/no-missing-import': ['error', {
           allowModules: ['license-badger']
         }]
       }
     },
     {
       extends: [
-        'ash-nazg/sauron-node-script-overrides',
+        'ash-nazg/sauron-node-overrides',
         'plugin:chai-expect/recommended',
         'plugin:chai-friendly/recommended'
       ],
@@ -42,7 +45,7 @@ module.exports = {
         mocha: true
       },
       parserOptions: {
-        ecmaVersion: 2018
+        ecmaVersion: 2022
       },
       rules: {
         'compat/compat': 0
@@ -54,9 +57,9 @@ module.exports = {
     'compat/compat': 0,
 
     // fs/promises
-    'node/no-missing-require': 0,
+    'n/no-missing-require': 0,
 
     // We need multiple exports
-    'node/exports-style': 0
+    'n/exports-style': 0
   }
 };