+++ src/glt-optionDefinitions.js
@@ -1,6 +1,10 @@
-'use strict';
+import {readFile} from 'fs/promises';
+import {dirname, join} from 'path';
+import {fileURLToPath} from 'url';
 
-const pkg = require('../package.json');
+const __dirname = dirname(fileURLToPath(import.meta.url));
+
+const pkg = JSON.parse(await readFile(join(__dirname, '../package.json')));
 
 // Todo: We really need a comamnd-line-args-TO-typedef-jsdoc generator!
 /* eslint-disable jsdoc/require-property -- See schema below */
@@ -29,5 +33,4 @@ const cliSections = [
   }
 ];
 
-exports.definitions = optionDefinitions;
-exports.sections = cliSections;
+export {optionDefinitions as definitions, cliSections as sections};